requirements.txt
python-binance == 0.7.9
schedule == 1.1.0
apprise == 0.9.1
matplotlib ==  3.3.3
numpy == 1.19.3
pandas == 1.2.0
pymongo == 3.11.3
dnspython == 2.1.0
openpyxl == 3.0.5
PyYAML == 5.4.1
websockets == 9.0.1
nest-asyncio == 1.5.1

